requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
google-generativeai==0.8.5
requests==2.32.3
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
protobuf==4.25.3
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-core==2.4.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
